Added multiple GTP Echo Requests and also received and printed the GTP Echo Response.

+++ src/testing/main.java
@@ -9,18 +9,23 @@ public class main {
 
 	public static void main(String[] args) 
 	{
-	
+
 		/** Delete logs from previous launches **/ 
 		File index = new File("logs");
 		String[]entries = index.list();
-		for(String s: entries){
-		    File currentFile = new File(index.getPath(),s);
-		    currentFile.delete();
+
+		if(entries != null)
+		{
+			for(String s: entries){
+				File currentFile = new File(index.getPath(),s);
+				currentFile.delete();
+			}
 		}
 		
 		SimulationControl.getInstance().startSimulation();
 		
-		
+
+
 	}
 
 }
